+++ addOns/ascanrulesAlpha/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+- Added hidden file scanner (issue 4585)
+
 ### Fixed
 - Fix typo in request header used by Apache Range Header DoS.